+++ brainpy/_src/tests/test_brainpy_deprecations.py
@@ -8,6 +8,8 @@ train_deprecated_names = list(brainpy.train.__deprecations.keys())
 dyn_deprecated_names = list(brainpy.dyn.__deprecations.keys())
 intg_deprecated_names = list(brainpy.integrators.__deprecations.keys())
 
+io_deprecated_names = list(brainpy.base.io.__deprecations.keys())
+
 
 class Test(parameterized.TestCase):
   @parameterized.product(
@@ -51,3 +53,10 @@ class Test(parameterized.TestCase):
   def test_brainpy_intg(self, name):
     with self.assertWarns(DeprecationWarning):
       getattr(brainpy.integrators, name)
+
+  @parameterized.product(
+    name=io_deprecated_names
+  )
+  def test_io(self, name):
+    with self.assertWarns(DeprecationWarning):
+      getattr(brainpy.base.io, name)
